# src/main/java/ru/bolodurin/socialmedia/model/dto/AuthResponse.java
package ru.bolodurin.socialmedia.model.dto;

import io.swagger.annotations.ApiModelProperty;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class AuthResponse {
    @ApiModelProperty(notes = "Bearer token")
    private String token;

    public static AuthResponse of(String token) {
        return new AuthResponse(token);
    }

}