# src/main/java/ru/bolodurin/socialmedia/model/dto/MessageRequest.java
package ru.bolodurin.socialmedia.model.dto;

import io.swagger.annotations.ApiModelProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class MessageRequest {
    @ApiModelProperty(notes = "Consumer username")
    private String consumer;
    @ApiModelProperty(notes = "Message content")
    private String message;

}